Further refactoring to implement check usage in rtcm3 tests
