--- c/tests/rtcm3_sbp_test.h
@@ -1,18 +1,0 @@
-/*
- * Copyright (C) 2017 Swift Navigation Inc.
- * Contact: Swift Navigation <dev@swiftnav.com>
- *
- * This source is subject to the license found in the file 'LICENSE' which must
- * be be distributed together with this source. All other rights reserved.
- *
- * THIS CODE AND INFORMATION IS PROVIDED "AS IS" WITHOUT WARRANTY OF ANY KIND,
- * EITHER EXPRESSED OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND/OR FITNESS FOR A PARTICULAR PURPOSE.
- */
-
-#ifndef GNSS_CONVERTERS_RTCM3_SBP_TEST_C_H
-#define GNSS_CONVERTERS_RTCM3_SBP_TEST_C_H
-
-int rtcm3_sbp_test(void);
-
-#endif // GNSS_CONVERTERS_RTCM3_SBP_TEST_C_H
